backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
SQLAlchemy==2.0.36
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0

